KIRO2023

Documentation for KIRO2023.

Parsing functions

This package exports the three following parsing functions:

read_instance: read an instance file from its path and returns a corresponding Instance object.

read_solution: write a Solution object to a solution format json file.

write_solution: read a solution file into a Solution object.

Data structures

Instance

Solution

Note: you may want to implement your own custom data structures that contain additional information that may be used by your algorithms.

Evaluation utilities

is_feasible: function that tests if a solution is feasible for a given instance

cost, operational_cost, construction_cost: functions that computes the cost of a solution for given instance

API

KIRO2023.Instance

KIRO2023.Solution

KIRO2023.compute_probabilities_of_failure

KIRO2023.construction_cost

KIRO2023.cost

KIRO2023.is_feasible

KIRO2023.operational_cost

KIRO2023.read_instance

KIRO2023.read_solution

KIRO2023.write_solution

KIRO2023.Instance — Type

Instance

Instance data structure.

KIRO2023.Solution — Type

Solution

Solution data structure.

KIRO2023.compute_probabilities_of_failure — Method

Return a probability of failure for each substation in solution

KIRO2023.construction_cost — Method

construction cost

KIRO2023.cost — Method

cost(solution::Solution, instance::Instance)

Compute the objective value of solution for given instance.

KIRO2023.is_feasible — Method

is_feasible(solution::Solution, instance::Instance; verbose=true)

Check if solution is feasible for instance. Prints some warnings when solution is infeasible and verbose is true.

KIRO2023.operational_cost — Method

operational cost

KIRO2023.read_instance — Method

read_instance(path::String)

Read instance from json file path.

KIRO2023.read_solution — Function

read_solution(path::String)

Read solution from json file path.

KIRO2023.write_solution — Method

write_solution(solution::Solution, path::String)

Write solution to file path with json format.